qt-qui: Refactor networking code
Make networking code use asynchronous signals etc. Although it turns out
that I needed to synchronize things with waitForBytesWritten and
waitForReadyRead anyway... The earlier to make sure commands are sent
right away, and the latter to not overflow the esp8266, which sometimes
crashes when it takes an influx of commands all at once (debug this
latter problem?).

Added code to parse status message and apply to sliders from server on
connect.

Sliders etc. are disabled by default until a connection is made.
Disabled whenever we get disconnected.

Master slider now starts out at max value.

TODO: box and button to connect to server with the GUI, rather than
having a hard-coded variant (also make it possible to specify on the
cmdline, however).