Refactor the server code out of VolumeControl and make it its own class
TCPVolumeServer is going to be the implementation of the old protocol.
Planning to add support for multiple clients at once using uselect.
Maybe implement feature where all clients are informed whenever there is
a state change in the volume controller.

UDPVolumeServer (currently a stub) is going to be a new connectionless
server protocol, which is actually a better fit for what we're trying to
do. Will only reply with the state of the VolumeController when
requested, in comparison with the TCP protocol. Clients will have to
poll the server for its state.