qt-gui: make GUI default to TCP protocol again, as UDP protocol had some issues with messages being ignored (which on status update would appear as the slider bouncing back from the position you set it to)
What likely is happing is that UDP messages sent while the server is
busy talking SPI or otherwise get ignored. To make this more robust we
should 1: make the server-side code faster if possible so less datagrams
are missed and 2: adorn the UDP protocol with sequence numbers, and the
server responding back with the ones it has received. The client will
retransmit the latest message, if it found out the latest message (where
the slider stopped) wasn't received. We are however not interested in
retransmitting old missed packages (which makes this not a case of
re-implementing TCP in UDP). SCTP would be nice, but micropython doesn't
support it...